Updated level, fixed gold counter

+++ mge_v17_student_version/src/Content/Hud/Hud.hpp
@@ -40,7 +40,7 @@ class Hud
 
 		sf::RenderWindow * _window;
 		int _coins = 0;
-		int _coinsDisplayed = 0;
+		float _coinsDisplayed = 0;
 
 		HUDMaterial* _hudMaterial;
         std::string _debugInfo;
